move number length estimation
